remove float32

+++ python/gps/algorithm/dynamics/dynamics_lr_prior.py
@@ -34,9 +34,9 @@ class DynamicsLRPrior(Dynamics):
         if N == 1:
             raise ValueError("Cannot fit dynamics on 1 sample")
 
-        self.Fm = np.zeros([T, dX, dX+dU]).astype(np.float32)
-        self.fv = np.zeros([T, dX]).astype(np.float32)
-        self.dyn_covar = np.zeros([T, dX, dX]).astype(np.float32)
+        self.Fm = np.zeros([T, dX, dX+dU])
+        self.fv = np.zeros([T, dX])
+        self.dyn_covar = np.zeros([T, dX, dX])
 
         it = slice(dX+dU)
         ip = slice(dX+dU, dX+dU+dX)
